fix: Add any to Props (because of error) #289
